backend/requirements.txt
datetime
fastapi 
uvicorn 
sqlalchemy 
aiopg 
psycopg2-binary 
prefect 
requests 
beautifulsoup4 
python-dotenv 
pydantic
pyquery